// src/api/get/getMemberById/types.ts
interface Role {
  id: number;
  name: string;
  guard_name: string;
  created_at: string;
  updated_at: string;
  nicename?: string;
  pivot: {
    model_type: string;
    model_id: number;
    role_id: number;
  };
}

interface Permission {
  id: number;
  name: string;
  guard_name: string;
  created_at: string;
  updated_at: string;
  pivot: {
    role_id: number;
    permission_id: number;
  };
}

export interface UserModel {
  id: number;
  name: string;
  email: string;
  email_verified_at: string | null;
  created_at: string;
  updated_at: string;
  phone?: string | null;
  position?: string | null;
  location?: string | null;
  avatar?: string | null;
  last_login?: string | null;
  last_ip?: string | null;
  roles?: Role[];
  permissions?: Permission[];
  can_register: boolean;
}